fix: rename service
